view select finish

+++ src/mvs/patch_match.h
@@ -141,11 +141,11 @@ struct PatchMatchOptions {
   bool write_consistency_graph = false;
 
   // Two-Stage Para
-  float pos_min_dis = 2;
-  float pos_max_dis = 8;
+  double pos_min_dis = 2;
+  double pos_max_dis = 8;
 
-  float ort_min_dis = 30;
-  float ort_max_dis = 135;
+  double ort_min_dis = 30;
+  double ort_max_dis = 135;
 
 
   void Print() const;
